style: split files to packages
