+++ app/src/main/res/layout/activity_student_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StudentListActivity">
+    tools:context=".screens.studentList.StudentListActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
